source-postgres: Move primary key FDB serialization logic
This belongs in datatypes.go with all the other datatype plumbing.
